Generate objects as QObject *
QML requires QObject's to be passed as pointers to instances. As such,
wrappers for managed objects will now be allocated in the heap. When no
longer needed, QML will delete wrappers that have a `nullptr` parent,
which should in turn guarantee that the corresponding .NET objects are
unpinned and made available for GC.

Change-Id: Ib37b50310f657b175a68df9bbdd8e7760aca80ad
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>