Rework event handling
Each event will now be mapped to just one signal with one parameter
corresponding to the EventArgs object, with the exception of property
change notifications that will also emit property notification signals.

This extends the EventArgs contract to QML code and is in keeping with the
goal of re-using native concepts as much as possible while avoiding the
use of Qt-specific native syntax (e.g. QSignal attribs in events).

The possibility to configure a custom conversion to signal parameters is
still a valid one, and might be re-introduced later on.

Change-Id: Iff91c40e32ff20bb6281f6f7ed709942a553fb06
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>