Add sourceRefs param to TestCodeGenerator
The `GenerateAsync` method can now receive a list of reference
assemblies to add to the ones already passed to the C# compilation step.

Also, the `System.Runtime` assembly is included as a default reference.
This is where several of the types exposed by the CoreLib are actually
implemented (e.g. `MulticastDelegate`, which is the underlying type for
`event` class members).

Change-Id: I2fd2cbdbcafaa356411dc308ce407810cfe3e553
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>