Add multi-thread support to QAIM
Model code between `Begin*` and `End*` calls will now execute in the
context of a critical section. Furthermore, the events that will trigger
the corresponding calls to the C++ functions `begin*` and `end*` are now
synchronized: the thread that sends one of these events will block until
it receives confirmation that the corresponding C++ call was made.

Change-Id: I83bd9bb975ea53e1549c3fdc2487a714c8a5a0cb
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>