Implement QML property semantics
QML will always call getter functions for object properties. This will
result in the creation of duplicate "live" C++ wrappers for the same C#,
one wrapper per reference to the property. To work around this, the
result of reference-typed properties (i.e. anything not considered
a built-in type) will be cached in the wrapper. Subsequent calls to the
property getter will return the cached instance. The cache is updated in
calls to the setter function (if any), and is cleared on notification of
property change.

Change-Id: I24dc08676c7c87093f0d8e77b775ec6dab9a8bfd
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>