Add DependsOn condition to generation rules
A rule that defines a `DependsOn` list of sources will not start execution
until all sources on the list have been successfully processed. If the
processing of a source is unsuccessful (i.e. one of the rules applied to
that source fails), all rules that depend on that source will fail.

Change-Id: Iee559ce5fe382067f03a93d6131c974684b269bc
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>